pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uk-reg-monitor"
version = "0.2.1"
description = "Regulatory monitoring client for the UK Employment Law Change Detector API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Desmond" },
]
keywords = ["uk", "regulation", "employment-law", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "APScheduler>=3.10,<4",
    "requests>=2.31",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "responses",
]

[project.scripts]
uk-reg-monitor = "uk_reg_monitor.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
